+++ include/Dictionary.hpp
